Refactoring PlanarAbility to interface of that name, and StdPlanarAbility as the new ability class id.
git-svn-id: svn://192.168.1.10/public/CoffeeMud@19330 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Libraries/CoffeeShops.java
@@ -249,7 +249,7 @@ public class CoffeeShops extends StdLibrary implements ShoppingLibrary
 						if(RawMaterial.CODES.VALUE(possCode) >= currVal)
 							valueChoices.add(Integer.valueOf(possCode));
 					}
-					material=(valueChoices.size()>0)?(valueChoices.get(CMath.abs(lieHash % valueChoices.size()))):I.material();
+					material=(valueChoices.size()>0)?(valueChoices.get(CMath.abs(lieHash % valueChoices.size()))).intValue():I.material();
 				}
 				else
 					material = I.material();
